+++ src/main/java/de/htwsaar/vs/chat/auth/jwt/JwtAuthenticationConverter.java
@@ -1,37 +1,34 @@
 package de.htwsaar.vs.chat.auth.jwt;
 
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.exceptions.JWTVerificationException;
-import de.htwsaar.vs.chat.util.ResponseError;
+import de.htwsaar.vs.chat.model.User;
+import org.springframework.core.ResolvableType;
+import org.springframework.core.codec.Decoder;
+import org.springframework.core.codec.Hints;
+import org.springframework.core.io.buffer.DataBuffer;
+import org.springframework.http.MediaType;
+import org.springframework.http.codec.json.Jackson2JsonDecoder;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.server.authentication.ServerAuthenticationConverter;
 import org.springframework.web.server.ServerWebExchange;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-
-import static com.auth0.jwt.algorithms.Algorithm.HMAC256;
-import static de.htwsaar.vs.chat.util.JwtUtil.JWT_PREFIX;
-import static de.htwsaar.vs.chat.util.JwtUtil.JWT_SECRET;
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
 public class JwtAuthenticationConverter implements ServerAuthenticationConverter {
 
-    @Override
-    public Mono<Authentication> convert(ServerWebExchange exchange) {
-        return Mono.justOrEmpty(exchange.getRequest().getHeaders().getFirst(AUTHORIZATION))
-                .filter(authorization -> authorization.toLowerCase().startsWith(JWT_PREFIX.toLowerCase()))
-                .map(authorization -> authorization.substring(JWT_PREFIX.length()))
-                .map(JwtAuthenticationConverter::verifyToken)
-                .onErrorResume(JWTVerificationException.class,
-                        e -> ResponseError.badRequest(e, "Could not verify JWT token"))
-                .map(username -> new UsernamePasswordAuthenticationToken(username, null));
+    private final Decoder<?> decoder;
+
+    public JwtAuthenticationConverter() {
+        decoder = new Jackson2JsonDecoder();
     }
 
-    private static String verifyToken(String token) {
-        return JWT.require(HMAC256(JWT_SECRET.getBytes()))
-                .build()
-                .verify(token)
-                .getClaim("name")
-                .asString();
+    @Override
+    public Mono<Authentication> convert(ServerWebExchange exchange) {
+        Flux<DataBuffer> body = exchange.getRequest().getBody();
+
+        return decoder
+                .decodeToMono(body, ResolvableType.forType(User.class), MediaType.APPLICATION_JSON, Hints.none())
+                .cast(User.class)
+                .map(u -> new UsernamePasswordAuthenticationToken(u.getUsername(), u.getPassword()));
     }
 }
